requirements.txt
neo4j>=5.22.0
pandas>=2.2.0
python-dotenv>=1.0.1
streamlit>=1.36.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
requests>=2.32.0
langgraph>=0.2.0
google-genai>=1.0.0
faker>=25.0.0
openai>=1.70.0
